Performance improvement by tweaking clocks
Patch by David F. Carlson to change the configuration settings for
SmartQ5/7 s3c6410 @ 666MHz with DDR333 Qimonda-HYB25DC512128CF-6

It *should* be stable but it is optimized for a particular spec of
memory, so it may fail if different components are used on other versions
of the board. Needs testing.